career_path_detector/requirements.txt
pandas>=2.0
pdfplumber>=0.10
pypdf>=3.0
python-docx>=1.0
google-generativeai>=0.5
